ml-api-master/requirements.txt
Flask==2.2.2
numpy==1.21.4
Pillow==9.2.0
tensorflow==2.7.0
google-cloud-storage==2.1.0
protobuf==3.20.*
Werkzeug==2.2.2